Split up button and icons for cart
