+++ assets/js/blocks/cart-checkout/cart/checkout-button/index.js
@@ -6,12 +6,12 @@ import { useState } from '@wordpress/element';
 import { Button } from '@woocommerce/base-components/cart-checkout';
 import { CHECKOUT_URL } from '@woocommerce/block-settings';
 import { useCheckoutContext } from '@woocommerce/base-context';
+
 /**
  * Internal dependencies
  */
-import { AmericanExpressLogo } from './american-express-logo';
 import './style.scss';
-import { MastercardLogo, VisaLogo } from './payment-logos'; // @todo we want to import this from `@automattic/composite-checkout` when it's published in NPM
+import PaymentMethodIcons from '../payment-method-icons';
 
 /**
  * Checkout button rendered in the full cart page.
@@ -31,11 +31,7 @@ const CheckoutButton = ( { link } ) => {
 			>
 				{ __( 'Proceed to Checkout', 'woo-gutenberg-products-block' ) }
 			</Button>
-			<div className="wc-block-cart__payment-methods">
-				<MastercardLogo />
-				<AmericanExpressLogo />
-				<VisaLogo />
-			</div>
+			<PaymentMethodIcons />
 		</div>
 	);
 };
