+++ assets/js/blocks/cart-checkout/cart/checkout-button/style.scss
@@ -1,0 +1,3 @@
+.wc-block-cart__submit-button {
+	width: 100%;
+}
