+++ assets/js/blocks/cart-checkout/cart/payment-method-icons/american-express-logo.js
