+++ assets/js/blocks/cart-checkout/cart/payment-method-icons/payment-logos.js
