+++ assets/js/blocks/cart-checkout/cart/payment-method-icons/style.scss
@@ -1,8 +1,4 @@
-.wc-block-cart__submit-button {
-	width: 100%;
-}
-
-.wc-block-cart__payment-methods {
+.wc-block-cart__payment-method-icons {
 	align-items: center;
 	display: flex;
 	justify-content: space-around;
